+++ features/settings/UsersPage.tsx
@@ -105,7 +105,6 @@ export const UsersPage: React.FC = () => {
 
     useEffect(() => {
         fetchUsers();
-        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
     useEffect(() => {